#78 Move interfaces related to state stores in new folder
Improve readability and faster to find interfaces
