+++ SD-K9/src/app/interfaces/states/building-info-state.d.ts
@@ -1,4 +1,4 @@
-import { OutdoorMapComponent } from "../components/outdoor-map/outdoor-map.component";
+import { OutdoorMapComponent } from "../../components/outdoor-map/outdoor-map.component";
 
 export declare interface BuildingInfoState {
     building: string;
