+++ SD-K9/src/app/interfaces/states/floor-plan-state.d.ts
@@ -1,4 +1,4 @@
-import { FloorPlanIdentifier } from './floor-plan-identifier';
+import { FloorPlanIdentifier } from '../floor-plan-identifier';
 
 export declare interface FloorPlanState {
     floorPlanIdentifiers: FloorPlanIdentifier[];
