+++ SD-K9/src/app/interfaces/states/google-map-state.d.ts
