+++ SD-K9/src/app/interfaces/states/map-mode-state.d.ts
