+++ SD-K9/src/app/interfaces/states/route-state.d.ts
@@ -1,4 +1,4 @@
-import { Route } from "./route";
+import { Route } from "../route";
 
 export declare interface RouteState {
     routes: Route[];
