+++ SD-K9/src/app/interfaces/states/state-reducer.d.ts
