+++ SD-K9/src/app/providers/state-stores/direction-form-store.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
-import { StateReducer } from '../../interfaces/state-reducer';
-import { StateAction } from '../../interfaces/state-action';
+import { StateReducer } from '../../interfaces/states/state-reducer';
+import { StateAction } from '../../interfaces/states/state-action';
 import { StateStore } from '../../helpers/state-store';
 import { DirectionForm } from '../../interfaces/direction-form';
 import { Action } from 'rxjs/internal/scheduler/Action';
